subchannel corresponds to one address
